+++ src/app/[lng]/template.tsx
@@ -1,17 +1,14 @@
 'use client';
-import { SENTRY_DSN } from '@/constants';
 import useDisableScrollBounce from '@/hooks/useDisableScrollBounce';
 import { useEasterEgg } from '@/hooks/useEasterEgg';
+import useListenMessageToReactNative from '@/hooks/useListenMessageToReactNative';
 import { useSentry } from '@/hooks/useSentry';
 
-export default function Step4Layout({ children }: { children: React.ReactNode }) {
+export default function Template({ children }: { children: React.ReactNode }) {
   useEasterEgg();
   useDisableScrollBounce();
-
-  useSentry({
-    dsn: SENTRY_DSN,
-    allowUrls: ['https://gloddy.vercel.app'],
-  });
+  useSentry();
+  useListenMessageToReactNative();
 
   return children;
 }
